Move serialiser library into own file
